move validation logic to the controller

+++ src/controllers/recipes.js
@@ -10,7 +10,15 @@ const getAll = async (req, res, next) => {
 
 const get = async (req, res, next) => {
   try {
-    res.json({ data: await service.get(req.params.id) });
+    const recipe = await service.get(req.params.id);
+
+    if (recipe === undefined) {
+      const err = new Error("Recipe not found");
+      err.statusCode = 404;
+      throw err;
+    }
+
+    res.json({ data: recipe });
   } catch (error) {
     next(error);
   }
@@ -42,6 +50,14 @@ const save = async (req, res, next) => {
 
 const update = async (req, res, next) => {
   try {
+    const recipe = await service.get(req.params.id);
+
+    if (recipe === undefined) {
+      const err = new Error("Recipe not found");
+      err.statusCode = 404;
+      throw err;
+    }
+
     const {
       name,
       healthLabels,
@@ -50,15 +66,15 @@ const update = async (req, res, next) => {
       ingredients,
     } = req.body;
 
-    const updatedRecipe = {
+    const updated = await service.update(req.params.id, {
       name,
       healthLabels: [...healthLabels],
       cookTimeMinutes,
       prepTimeMinutes,
       ingredients: [...ingredients],
-    };
+    });
 
-    res.json({ data: await service.update(req.params.id, updatedRecipe) });
+    res.json({ data: updated });
   } catch (error) {
     next(error);
   }
@@ -66,6 +82,14 @@ const update = async (req, res, next) => {
 
 const remove = async (req, res, next) => {
   try {
+    const recipe = await service.get(req.params.id);
+
+    if (recipe === undefined) {
+      const err = new Error("Recipe not found");
+      err.statusCode = 404;
+      throw err;
+    }
+
     await service.remove(req.params.id);
     res.sendStatus(204);
   } catch (error) {
